3161: Moved the MenuItem to ChatMenu and adjusted the Dialog width
