+++ web/src/components/ChatController.tsx
@@ -22,7 +22,6 @@ type ChatControllerProps = {
   updateChatId: (newChatId: string | null) => void
 }
 
-export const LOCAL_STORAGE_ITEM_CHAT_MESSAGES = 'Chat-Device-Id'
 const LOCAL_STORAGE_ITEM_CHAT_PRIVACY_POLICIES = 'Chat-Privacy-Policies'
 const DEFAULT_POLLING_INTERVAL = 15000
 const TYPING_POLLING_INTERVAL = 3000
