+++ web/src/components/ChatMenu.tsx
@@ -1,34 +1,35 @@
+import AddCommentOutlinedIcon from '@mui/icons-material/AddCommentOutlined'
 import MoreVertIcon from '@mui/icons-material/MoreVert'
 import Button from '@mui/material/Button'
 import DialogActions from '@mui/material/DialogActions'
 import IconButton from '@mui/material/IconButton'
 import MuiMenu from '@mui/material/Menu'
+import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
-import { styled } from '@mui/material/styles'
 import React, { ReactElement, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
+import MenuItem from './MenuItem'
 import AlertDialog from './base/AlertDialog'
 
-const StyledDialogContainer = styled('div')({
-  display: 'flex',
-  flexDirection: 'column',
-})
-
 type ChatMenuProps = {
-  confirmNewChatOpen: boolean
-  onConfirmClose: () => void
-  onConfirmNewChat: () => void
-  children: ReactElement | ReactElement[]
+  chatId: string | null
+  updateChatId: (newValue: string | null) => void
 }
 
-const ChatMenu = ({ confirmNewChatOpen, onConfirmClose, onConfirmNewChat, children }: ChatMenuProps): ReactElement => {
+const ChatMenu = ({ chatId, updateChatId }: ChatMenuProps): ReactElement => {
   const [menuAnchorElement, setMenuAnchorElement] = useState<HTMLElement | null>(null)
+  const [confirmDialogOpen, setConfirmDialogOpen] = useState(false)
   const { t } = useTranslation('chat')
 
   const handleDialogAction = (action: () => void) => {
     setMenuAnchorElement(null)
     action()
+  }
+
+  const resetChatId = () => {
+    updateChatId(null)
+    setConfirmDialogOpen(false)
   }
 
   return (
@@ -41,28 +42,36 @@ const ChatMenu = ({ confirmNewChatOpen, onConfirmClose, onConfirmNewChat, childr
       </IconButton>
       <MuiMenu
         anchorEl={menuAnchorElement}
-        open={menuAnchorElement !== null && !confirmNewChatOpen}
+        open={menuAnchorElement !== null && !confirmDialogOpen}
         onClose={() => setMenuAnchorElement(null)}>
-        {children}
+        <MenuItem
+          text={t('newChat')}
+          icon={<AddCommentOutlinedIcon fontSize='small' />}
+          disabled={chatId === null}
+          onClick={() => setConfirmDialogOpen(true)}
+        />
       </MuiMenu>
-      {confirmNewChatOpen && (
+      {confirmDialogOpen && (
         <AlertDialog
           title={t('newChat')}
-          close={onConfirmClose}
+          close={() => setConfirmDialogOpen(false)}
           actions={
             <DialogActions>
-              <Button onClick={() => handleDialogAction(onConfirmClose)} variant='outlined' sx={{ flex: '1 1' }}>
+              <Button
+                onClick={() => handleDialogAction(() => setConfirmDialogOpen(false))}
+                variant='outlined'
+                sx={{ flex: '1 1' }}>
                 {t('layout:cancel')}
               </Button>
-              <Button onClick={() => handleDialogAction(onConfirmNewChat)} variant='contained' sx={{ flex: '3 3' }}>
+              <Button onClick={() => handleDialogAction(resetChatId)} variant='contained' sx={{ flex: '3 3' }}>
                 {t('newChat')}
               </Button>
             </DialogActions>
           }>
-          <StyledDialogContainer>
-            <Typography variant='body2'>{t('newChatConfirmationTitle')}</Typography>
+          <Stack>
+            <Typography variant='body2'>{t('newChatConfirmation')}</Typography>
             <Typography variant='body2'>{t('newChatConfirmationMessage')}</Typography>
-          </StyledDialogContainer>
+          </Stack>
         </AlertDialog>
       )}
     </>
